+++ src/components/formButtons/PauseFormButton.tsx
@@ -1,4 +1,7 @@
+import { Breakpoint, useBreakpoint } from '@helsenorge/designsystem-react';
 import Button from '@helsenorge/designsystem-react/components/Button';
+import Icon from '@helsenorge/designsystem-react/components/Icon';
+import Save from '@helsenorge/designsystem-react/components/Icons/Save';
 
 type Props = {
   onPauseButtonClicked?: () => void;
@@ -7,13 +10,20 @@ type Props = {
   isStepView?: boolean;
 };
 
-export const PauseFormButton = ({ pauseButtonText, onPauseButtonClicked, pauseButtonDisabled }: Props): JSX.Element | null => {
+export const PauseFormButton = ({ pauseButtonText, onPauseButtonClicked, pauseButtonDisabled, isStepView }: Props): JSX.Element | null => {
   if (!onPauseButtonClicked) {
     return null;
   }
+
+  const breakpoint = useBreakpoint();
+
+  const displaySaveIcon = breakpoint > Breakpoint.xxs && !isStepView;
+
+  useBreakpoint;
   return (
     <div>
       <Button variant="outline" disabled={pauseButtonDisabled} onClick={onPauseButtonClicked} testId="refero-pause-button">
+        {displaySaveIcon && <Icon svgIcon={Save} />}
         {pauseButtonText}
       </Button>
     </div>