+++ src/components/formcomponents/attachment/attachmenthtml.tsx
@@ -1,6 +1,4 @@
-import { useState } from 'react';
-
-import { Attachment, QuestionnaireItem } from 'fhir/r4';
+import { QuestionnaireItem } from 'fhir/r4';
 import { FieldValues, RegisterOptions, useFormContext } from 'react-hook-form';
 import styles from '../common-styles.module.css';
 import FormGroup from '@helsenorge/designsystem-react/components/FormGroup';
@@ -9,7 +7,12 @@ import NotificationPanel from '@helsenorge/designsystem-react/components/Notific
 import FileUpload, { MimeTypes, UploadFile } from '@helsenorge/file-upload/components/file-upload';
 import { useFileUpload } from '@helsenorge/file-upload/components/file-upload/useFileUpload';
 
-import { validateFileSize, validateFileType, validateMaxFiles, validateMinFiles } from './attachment-validation';
+import { getCustomValidationText } from './attachment-validation';
+import {
+  validateNumberOfFiles,
+  validateFileType,
+  validateTotalFileSize,
+} from '@helsenorge/file-upload/components/file-upload/validate-utils';
 import { getAttachmentMaxSizeBytesToUse } from './attachmentUtil';
 import { VALID_FILE_TYPES } from '@/constants';
 import { getId, isReadOnly } from '@/util';
@@ -17,11 +20,12 @@ import { Resources } from '@/util/resources';
 
 import { ReferoLabel } from '@/components/referoLabel/ReferoLabel';
 import { TextMessage } from '@/types/text-message';
-import { useGetAnswer } from '@/hooks/useGetAnswer';
 import { QuestionnaireComponentItemProps } from '@/components/createQuestionnaire/GenerateQuestionnaireComponents';
 import { ReadOnly } from '../read-only/readOnly';
 import { getErrorMessage, required } from '@/components/validation/rules';
 import { shouldValidate } from '@/components/validation/utils';
+import { useAttachmentSync } from './useAttachmentSync';
+import { useEffect } from 'react';
 
 type Props = QuestionnaireComponentItemProps & {
   onUpload: (files: UploadFile[]) => void;
@@ -40,7 +44,7 @@ type Props = QuestionnaireComponentItemProps & {
   minFiles?: number;
   item?: QuestionnaireItem;
   attachmentMaxFileSize?: number;
-  attachmentValidTypes?: Array<string>;
+  attachmentValidTypes?: MimeTypes[];
   idWithLinkIdAndItemIndex: string;
   customErrorMessage: TextMessage | undefined;
 };
@@ -66,86 +70,80 @@ const AttachmentHtml = (props: Props): JSX.Element | null => {
     idWithLinkIdAndItemIndex,
     customErrorMessage,
     path,
-    linkId,
     pdf,
     children,
     multiple,
+    onRequestAttachmentLink,
   } = props;
 
   const { formState, getFieldState, register: internalRegister } = useFormContext<FieldValues>();
-  const [disableButton, setDisableButton] = useState(false);
+
   const fieldState = getFieldState(idWithLinkIdAndItemIndex, formState);
   const { error } = fieldState;
 
-  const answer = useGetAnswer(linkId, path);
-  const getMaxValueBytes = getAttachmentMaxSizeBytesToUse(attachmentMaxFileSize, item);
-  const validFileTypes = attachmentValidTypes ? attachmentValidTypes : VALID_FILE_TYPES;
-  const deleteText = resources ? resources.deleteAttachmentText : undefined;
-  const { acceptedFiles, rejectedFiles, setAcceptedFiles, setRejectedFiles, register } = useFileUpload(
+  const numberOfFilesMessage = getCustomValidationText(
+    item,
+    (minFiles && !maxFiles && resources?.attachmentError_minFiles) ||
+      (maxFiles && !minFiles && resources?.attachmentError_maxFiles) ||
+      (maxFiles && minFiles && resources?.attachmentError_minFiles) ||
+      undefined
+  );
+  const validationFileTypesMessage = getCustomValidationText(item, resources?.attachmentError_fileType);
+  const maxValueBytes = getAttachmentMaxSizeBytesToUse(attachmentMaxFileSize, item);
+  const validFileTypes: MimeTypes[] = attachmentValidTypes ? attachmentValidTypes : VALID_FILE_TYPES;
+
+  const {
+    setAcceptedFiles,
+    setRejectedFiles,
+    register,
+    acceptedFiles: extAccepted,
+    rejectedFiles: extRejected,
+  } = useFileUpload(
     internalRegister,
+    [validateFileType(validFileTypes, validationFileTypesMessage)],
     [
-      (file): true | string => (file ? validateFileType(file, validFileTypes, item, resources?.attachmentError_fileType) : true),
-      (file): true | string => (file ? validateFileSize(file, getMaxValueBytes, item, resources?.attachmentError_fileSize) : true),
-    ],
-    [
-      (files): true | string =>
-        files.length && minFiles ? validateMinFiles(files, minFiles, item, resources?.attachmentError_minFiles) : true,
-      (files): true | string =>
-        files.length && maxFiles ? validateMaxFiles(files, maxFiles, item, resources?.attachmentError_maxFiles) : true,
+      validateNumberOfFiles(minFiles ?? 0, maxFiles ?? 20, numberOfFilesMessage || 'Number of files'),
+      validateTotalFileSize(0, maxValueBytes, getCustomValidationText(item, resources?.attachmentError_fileSize || 'tiotal file size')),
     ]
   );
 
-  const handleUpload = (files: UploadFile[]): void => {
-    onUpload(files);
-    setAcceptedFiles(prevState => [...prevState, ...files]);
-    if (!multiple) {
-      setDisableButton(true);
+  const { acceptedFiles, rejectedFiles, handleDelete, handleUpload, disableButton, value } = useAttachmentSync({
+    onUpload,
+    path,
+    onDelete,
+    rejectedFiles: extRejected,
+    acceptedFiles: extAccepted,
+    setAcceptedFiles,
+    setRejectedFiles,
+    item,
+    idWithLinkIdAndItemIndex,
+  });
+
+  useEffect(() => {
+    if (onRequestAttachmentLink && acceptedFiles.length > 0) {
+      for (const file of acceptedFiles) {
+        onRequestAttachmentLink(file.id);
+      }
     }
-  };
+  }, [acceptedFiles, onRequestAttachmentLink]);
 
-  const handleDelete = (fileId: string): void => {
-    onDelete(fileId);
-    setAcceptedFiles(acceptedFiles.filter(x => x.id !== fileId));
-    setRejectedFiles(rejectedFiles.filter(x => x.id !== fileId));
-    if (!multiple) {
-      setDisableButton(false);
-    }
-  };
-
-  const getAttachmentValue = (): Attachment | Attachment[] | undefined => {
-    if (Array.isArray(answer)) {
-      return answer.map(v => v.valueAttachment).filter((attachment): attachment is Attachment => attachment !== undefined);
-    } else if (answer && answer.valueAttachment) {
-      return answer.valueAttachment;
-    }
-    return undefined;
-  };
-
-  const attachmentValue = getAttachmentValue();
-
-  const getAttachmentValueForPdf = (): UploadedFile[] | undefined => {
-    if (Array.isArray(attachmentValue)) {
-      return attachmentValue.map(attachment => {
-        return {
-          id: attachment.url ?? '-1',
-          name: attachment.title || '',
-        };
-      });
-    } else if (attachmentValue && attachmentValue.url) {
-      return [
-        {
-          id: attachmentValue.url,
-          name: attachmentValue.title || '',
-        },
-      ];
-    }
-
-    return undefined;
-  };
+  const deleteText = resources ? resources.deleteAttachmentText : undefined;
 
   const getPdfValue = (): string => {
+    const getAttachmentValueForPdf = (): UploadedFile[] | undefined => {
+      if (Array.isArray(value)) {
+        return value.map(attachment => {
+          return {
+            id: attachment.url ?? '-1',
+            name: attachment.title || '',
+          };
+        });
+      }
+
+      return undefined;
+    };
     const attachmentValueForPdf = getAttachmentValueForPdf();
-    if (attachmentValueForPdf) {
+    if (attachmentValueForPdf && attachmentValueForPdf.length > 0) {
       return attachmentValueForPdf.map(v => v.name).join(', ');
     } else if (resources) {
       return resources.ikkeBesvart;
@@ -159,8 +157,6 @@ const AttachmentHtml = (props: Props): JSX.Element | null => {
     validate: () => true,
     shouldUnregister: true,
   };
-  const { onChange, ...rest } = register(idWithLinkIdAndItemIndex, shouldValidate(item, pdf) ? validationRules : undefined);
-
   if (pdf || isReadOnly(item)) {
     return (
       <ReadOnly
@@ -168,7 +164,7 @@ const AttachmentHtml = (props: Props): JSX.Element | null => {
         id={id}
         idWithLinkIdAndItemIndex={idWithLinkIdAndItemIndex}
         item={item}
-        value={attachmentValue}
+        value={value}
         pdfValue={getPdfValue()}
         errors={error}
       >
@@ -189,12 +185,9 @@ const AttachmentHtml = (props: Props): JSX.Element | null => {
           attachmentLabel={resources?.supportedFileFormats}
         />
         <FileUpload
-          {...rest}
+          {...register(idWithLinkIdAndItemIndex, shouldValidate(item, pdf) ? validationRules : undefined)}
           wrapperTestId={`${getId(id)}-attachment`}
           inputId={id}
-          onChange={e => {
-            onChange(e);
-          }}
           onChangeFile={handleUpload}
           onDeleteFile={handleDelete}
           chooseFilesText={resources?.chooseFilesText}