+++ src/components/formcomponents/quantity/quantity.tsx
@@ -6,7 +6,7 @@ import Input from '@helsenorge/designsystem-react/components/Input';
 import styles from './quantity.module.css';
 import { newQuantityValueAsync } from '@/actions/newValue';
 import { GlobalState, useAppDispatch } from '@/reducers';
-import { getMaxValueExtensionValue, getPlaceholder, getQuestionnaireUnitExtensionValue } from '@/util/extension';
+import { getMaxValueExtensionValue, getMinValueExtensionValue, getPlaceholder, getQuestionnaireUnitExtensionValue } from '@/util/extension';
 import { isReadOnly, getId } from '@/util/index';
 
 import { ReferoLabel } from '@/components/referoLabel/ReferoLabel';
@@ -16,11 +16,12 @@ import RenderDeleteButton from '../repeat/RenderDeleteButton';
 import RenderRepeatButton from '../repeat/RenderRepeatButton';
 import { QuestionnaireComponentItemProps } from '@/components/createQuestionnaire/GenerateQuestionnaireComponents';
 import { useExternalRenderContext } from '@/context/externalRenderContext';
-import { decimalPattern, getErrorMessage, maxValue, minValue, required } from '@/components/validation/rules';
+import { decimalPattern, getErrorMessage, getInputWidth, maxValue, minValue, required } from '@/components/validation/rules';
 import { findQuestionnaireItem } from '@/reducers/selectors';
 import useOnAnswerChange from '@/hooks/useOnAnswerChange';
 import { ReadOnly } from '../read-only/readOnly';
 import { shouldValidate } from '@/components/validation/utils';
+import { useResetFormField } from '@/hooks/useResetFormField';
 
 export type Props = QuestionnaireComponentItemProps;
 
@@ -47,7 +48,9 @@ const Quantity = (props: Props): JSX.Element | null => {
       return answer.valueQuantity.value;
     }
   };
+  const value = getValue(answer);
 
+  useResetFormField(idWithLinkIdAndItemIndex, value);
   const getPDFValue = (): string => {
     const value = getValue(answer);
     if (value === undefined || value === null) {
@@ -94,11 +97,10 @@ const Quantity = (props: Props): JSX.Element | null => {
     return '';
   };
 
-  const value = getValue(answer);
   const maxCharacters = getMaxValueExtensionValue(item) ? getMaxValueExtensionValue(item)?.toString().length : undefined;
-  const width = maxCharacters ? (maxCharacters > 40 ? 40 : maxCharacters + 2) : 7;
+  const baseIncrementValue = getMinValueExtensionValue(item);
+  const width = getInputWidth(maxCharacters);
   const errorMessage = getErrorMessage(item, error);
-
   const validationRules: RegisterOptions<FieldValues, string> | undefined = {
     required: required({ item, resources }),
     max: maxValue({ item, resources }),
@@ -150,6 +152,7 @@ const Quantity = (props: Props): JSX.Element | null => {
               handleChange(e);
             }}
             width={width}
+            baseIncrementValue={baseIncrementValue}
           />
           <span className={`${styles.pageReferoUnit} page_refero__unit`}>{getUnit()}</span>
         </div>