+++ src/reducers/form.ts
@@ -53,10 +53,12 @@ import { createSlice, current, PayloadAction } from '@reduxjs/toolkit';
 
 export interface FormData {
   Content: QuestionnaireResponse | null | undefined;
+  isExternalUpdate?: boolean;
 }
 
 export interface FormDefinition {
   Content: Questionnaire | null | undefined;
+  isExternalUpdate?: boolean;
 }
 
 export interface Form {
@@ -76,13 +78,18 @@ const initialState: Form = {
   },
   FormDefinition: {
     Content: null,
+    isExternalUpdate: false,
   },
   Language: LanguageLocales.NORWEGIAN.toLowerCase(),
 };
 const formSlice = createSlice({
   name: 'form',
   initialState,
-  reducers: {},
+  reducers: {
+    setIsExternalUpdateAction(state, action: PayloadAction<boolean>) {
+      state.FormData.isExternalUpdate = action.payload;
+    },
+  },
   extraReducers: builder => {
     builder
       .addCase(setSkjemaDefinitionAction, (state, action: PayloadAction<SetFormDefinitionAction>) => {
@@ -112,6 +119,7 @@ const formSlice = createSlice({
   },
 });
 export default formSlice.reducer;
+export const actions = formSlice.actions;
 
 export function getFormData(state: GlobalState): FormData | null {
   if (!state.refero?.form?.FormData) {
@@ -410,8 +418,7 @@ function processRemoveAttachmentValueAction(action: NewValuePayload, state: Form
   }
 
   if (action.valueAttachment) {
-    const attachmentToRemove = action.valueAttachment.title;
-    var index = responseItem.answer.findIndex(el => el && el.valueAttachment && el.valueAttachment.title === attachmentToRemove);
+    const index = responseItem.answer.findIndex(el => el && el.valueAttachment && el.valueAttachment.id === action.valueAttachment?.id);
     if (index > -1) {
       responseItem.answer.splice(index, 1);
     }
@@ -502,6 +509,7 @@ function processNewValueAction(payload: NewValuePayload, state: Form): Form {
     hasAnswer = true;
 
     const attachment: Attachment = {
+      id: payload.valueAttachment.id,
       url: payload.valueAttachment.url,
       title: payload.valueAttachment.title,
       data: payload.valueAttachment.data,
@@ -834,17 +842,20 @@ function processSetSkjemaDefinition(payload: SetFormDefinitionAction, state: For
     Content: payload.questionnaire,
   };
   const statetest = current(state.FormData);
+
   let formData: FormData;
+
   if (payload.questionnaireResponse && payload.syncQuestionnaireResponse) {
-    formData = { Content: syncQuestionnaireResponse(payload.questionnaire, payload.questionnaireResponse) };
+    formData = { Content: syncQuestionnaireResponse(payload.questionnaire, payload.questionnaireResponse), isExternalUpdate: true };
   } else if (payload.questionnaireResponse) {
-    formData = { Content: payload.questionnaireResponse };
+    formData = { Content: payload.questionnaireResponse, isExternalUpdate: true };
   } else if (statetest?.Content === initialState.FormData.Content) {
     formData = { Content: generateQuestionnaireResponse(payload.questionnaire) };
   } else {
     formData = state.FormData;
   }
   state.FormDefinition = formDefinition;
+
   state.FormData = formData;
   state.Language = payload.language || state.Language;
   return state;